+++ src/callbacks/project_page/callback_tabs_switch_project_page.py
@@ -63,5 +63,5 @@ def update_project_page_visualization(tabs_switch, result_type: str, imported_ru
 
     map_fig = plot_project_overview_map(projects, trajects.values())
     cost, damage, risk = calc_area_stats(projects)
-    return cost_fig, reliability_fig, map_fig, project_overview_table, f"{cost/1e6:.2f} M€", f"{damage/1e6:.2f} M€", f"{risk:.2f} M€"
+    return cost_fig, reliability_fig, map_fig, project_overview_table, f"{cost/1e6:.2f} M€", f"{damage/1e6:.2f} M€", f"{risk:.2f} €"